Update simss.pas
If the HTL faces the illumination, with the previous version one cannot use "untilVoc = 1". When using "untilVoc = 1" and "Vscan = -1" in the previous version, at the first point Jext is positive, because the JV curve should be drawn from the 2nd to the 3rd quadrant. Therefore SIMsalabim quits immediately after the first point.

Note that with this change, Vmpp, MPP, and FF are not correctly determined in the scPars-file for if the HTL faces the illumination. This need further improvement.